Extension blocks + image data
Both the Graphic Control Extension block and Comment Extension blocks
are now recognized and parsed. The former currently throws away the data
extracted because it is currently not relevant to our rendering (it'll
be relevant as soon as transparent images are encountered), and the
latter just has its contents printed, then thrown away.

The Image Descriptor, local color tables and the image data are now all
parsed and extracted (technically, no local color table has been tested,
and the test GIF only contains one image).

The next step is to perform LZW decoding!