First image successfully displayed!
Success! gfx/rgb-stripes.gif is sucessfully displayed in its own window.
The two immediate issues to fix is that the active color table is
currently hard coded to be the global color table, and the transparency
is not actually implemented (though gfx/rgb-stripes-transparent.gif
works fine otherwise). Both of these issues have to do with the fact
that some of the data parsed is not being kept around in the final data
structure, and so is not available for use at display time.

I checked a bunch of other GIF files, but either they don't work
(failing during or before parsing), or in the case of one, shows up
rotated for some reason.

Also, it would be nice to have a zoom feature...