Added option to enter filename when executing
If the erc-compiled executable is run through erl directly, the user
must pass in the filename to load. The Makefile has also been modified
to pass in a default filename when executing the "run" target. The
filename can be overridden by setting the environment variable IN to the
desired value, e.g. "make run IN=gfx/rgb-stripes.gif".