Implement variable-length code reading + tests
This version of the code reading is much less hacky than the "convert a
bitstream into a giant string" approach used earlier. Furthermore, with
the corresponding tests, I have a higher confidence in the correctness
of the reading process.

The new code is the first example of EUnit tests, as well as the use of
the EDoc comment syntax.

This new algorithm is not yet used.