Rewrite decompression algorithm
With the new code reader, I've re-written the entire decompression
algorithm. Not only does the algorithm run to completion for more
complicated test images, but does so reasonably fast. I took this
opportunity to better document the code, as well as make it easier to
understand with better naming and structure.

There still seems to be some issues related to how the parsed image is
displayed. I'm not sure if these issues are due to the actual
decompression, or due to the rendering logic.