Ignore arbitrary extension blocks
This commit supports reading, then discarding, unhandled extension
blocks. In particular, this is important for rudimentary animated GIF
support, which requires parsing the Application Extension block. This
block contains meta-data about the animation, such as the number of
times to loop, so for now, I'm able to ignore this block. However, I do
need to ensure I can read this block so I can ignore it!

Extension block data, like the image data, is grouped into sub-blocks.
For that reason, I refactored the `image_data_next_sub_block/2` function
so I could use it for extensions as well.

Technically, the comment extension is also broken into sub-blocks. I was
only reading the first sub-block, with the assumption that there was
only one. This assumption breaks when the comment is more than 255 bytes
long! Thus, this commit fixes that bug as well.