Allow graphics control to target later images
As per section 23.e. of the GIF89A spec:

> The scope of this Extension is the graphic rendering block that
> follows it; it is possible for other extensions to be present between
> this block and its target.

This is despite the official grammar stating that a graphic control
extension block, if present, is followed immediately by a graphic
rendering block (either an image or a plain text extension).

As I've encountered real-world GIFs with a comment extension in between
a graphic rendering block and an image, I've modified the parser to
allow such configurations.